Summary Arithmetic overflow fix
Fix error generating summary if min estimated days exceeds int max value
#990

+++ DBADashDB/dbo/Tables/Summary.sql
@@ -60,6 +60,6 @@
     DatabaseStateStatus INT NOT NULL,
 	RefreshDate DATETIME NOT NULL,
 	IsAzure BIT NOT NULL,
-    MinIdentityEstimatedDays INT NULL,
+    MinIdentityEstimatedDays BIGINT NULL,
 	CONSTRAINT PK_Summary PRIMARY KEY(InstanceID)
 );
